it actually seems to work now

+++ Protocol/protocol.cpp
@@ -1,3 +1,3 @@
 #include "protocol.hpp"
 
-const int DPProtocol::c_magicNumber[] = {0x44, 0x50};
+const uint8_t DPProtocol::c_magicNumber[] = {0x44, 0x50};
